Renames useNormalizedCMSResults to avoid confusion concerning hook naming conventions
